Implement: Connection procedure + timeout checking
the combination of PASS NICK USER is now needed to connect to the server.
Pass can be tried many time but no feedback is given and only the last try is kept
Nick can also be given many times, it will change the registered nickname.
User can be supplied only once , following tries will produce an already registered error
If all thoses steps are not performed and validated before CONNECTION_TIMEOUT, the user is destroyed and the connection closed